type payload in workflow-execution-event- and agent-execution-event-
