+++ packages/core/src/stream/types.ts
@@ -16,7 +16,7 @@ import type { AIV5Type } from '../agent/message-list/types';
 import type { StructuredOutputOptions } from '../agent/types';
 import type { TracingContext } from '../ai-tracing/types';
 import type { OutputProcessor } from '../processors';
-import type { WorkflowStreamEvent } from '../workflows/types';
+import type { WorkflowRunStatus, WorkflowStepStatus } from '../workflows/types';
 import type { InferSchemaOutput, OutputSchema, PartialSchemaOutput } from './base/schema';
 
 export enum ChunkFrom {
@@ -458,11 +458,11 @@ export type NetworkChunkType =
   | (BaseChunkType & { type: 'tool-execution-end'; payload: ToolExecutionEndPayload })
   | (BaseChunkType & { type: 'network-execution-event-step-finish'; payload: NetworkStepFinishPayload })
   | (BaseChunkType & { type: 'network-execution-event-finish'; payload: NetworkFinishPayload })
-  | (BaseChunkType & { type: `agent-execution-event-${string}`; payload: unknown })
-  | (BaseChunkType & { type: `workflow-execution-event-${string}`; payload: object });
+  | (BaseChunkType & { type: `agent-execution-event-${string}`; payload: AgentChunkType })
+  | (BaseChunkType & { type: `workflow-execution-event-${string}`; payload: WorkflowStreamEvent });
 
 // Strongly typed chunk type (currently only OUTPUT is strongly typed, tools use dynamic types)
-export type TypedChunkType<OUTPUT extends OutputSchema = undefined> =
+export type AgentChunkType<OUTPUT extends OutputSchema = undefined> =
   | (BaseChunkType & { type: 'response-metadata'; payload: ResponseMetadataPayload })
   | (BaseChunkType & { type: 'text-start'; payload: TextStartPayload })
   | (BaseChunkType & { type: 'text-delta'; payload: TextDeltaPayload })
@@ -502,10 +502,92 @@ export type TypedChunkType<OUTPUT extends OutputSchema = undefined> =
     })
   | (BaseChunkType & { type: 'tool-output'; payload: DynamicToolOutputPayload })
   | (BaseChunkType & { type: 'step-output'; payload: StepOutputPayload })
-  | (BaseChunkType & { type: 'workflow-step-output'; payload: StepOutputPayload })
   | (BaseChunkType & { type: 'watch'; payload: WatchPayload })
-  | (BaseChunkType & { type: 'tripwire'; payload: TripwirePayload })
-  | (BaseChunkType & WorkflowStreamEvent)
+  | (BaseChunkType & { type: 'tripwire'; payload: TripwirePayload });
+
+export type WorkflowStreamEvent =
+  | (BaseChunkType & {
+      type: 'workflow-start';
+      payload: {
+        workflowId: string;
+      };
+    })
+  | (BaseChunkType & {
+      type: 'workflow-finish';
+      payload: {
+        workflowStatus: WorkflowRunStatus;
+        output: {
+          usage: {
+            inputTokens: number;
+            outputTokens: number;
+            totalTokens: number;
+          };
+        };
+        metadata: Record<string, any>;
+      };
+    })
+  | (BaseChunkType & {
+      type: 'workflow-canceled';
+      payload: {};
+    })
+  | (BaseChunkType & {
+      type: 'workflow-step-start';
+      id: string;
+      payload: {
+        id: string;
+        stepCallId: string;
+        status: WorkflowStepStatus;
+        output?: Record<string, any>;
+        payload?: Record<string, any>;
+        resumePayload?: Record<string, any>;
+        suspendPayload?: Record<string, any>;
+      };
+    })
+  | (BaseChunkType & {
+      type: 'workflow-step-finish';
+      payload: {
+        id: string;
+        metadata: Record<string, any>;
+      };
+    })
+  | (BaseChunkType & {
+      type: 'workflow-step-suspended';
+      payload: {
+        id: string;
+        status: WorkflowStepStatus;
+        output?: Record<string, any>;
+        payload?: Record<string, any>;
+        resumePayload?: Record<string, any>;
+        suspendPayload?: Record<string, any>;
+      };
+    })
+  | (BaseChunkType & {
+      type: 'workflow-step-waiting';
+      payload: {
+        id: string;
+        payload: Record<string, any>;
+        startedAt: number;
+        status: WorkflowStepStatus;
+      };
+    })
+  | (BaseChunkType & { type: 'workflow-step-output'; payload: StepOutputPayload })
+  | (BaseChunkType & {
+      type: 'workflow-step-result';
+      payload: {
+        id: string;
+        stepCallId: string;
+        status: WorkflowStepStatus;
+        output?: Record<string, any>;
+        payload?: Record<string, any>;
+        resumePayload?: Record<string, any>;
+        suspendPayload?: Record<string, any>;
+      };
+    });
+
+// Strongly typed chunk type (currently only OUTPUT is strongly typed, tools use dynamic types)
+export type TypedChunkType<OUTPUT extends OutputSchema = undefined> =
+  | AgentChunkType<OUTPUT>
+  | WorkflowStreamEvent
   | NetworkChunkType;
 
 // Default ChunkType for backward compatibility using dynamic (any) tool types
