+++ packages/core/src/workflows/types.ts
@@ -4,8 +4,9 @@ import type { TracingPolicy, TracingProperties } from '../ai-tracing';
 import type { Mastra } from '../mastra';
 import type { ExecutionEngine } from './execution-engine';
 import type { ConditionFunction, ExecuteFunction, LoopConditionFunction, Step } from './step';
+import type { WorkflowStreamEvent } from '../stream/types';
 
-export type { ChunkType } from '../stream/types';
+export type { ChunkType, WorkflowStreamEvent } from '../stream/types';
 export type { MastraWorkflowStream } from '../stream/MastraWorkflowStream';
 
 export type Emitter = {
@@ -145,84 +146,6 @@ export type StreamEvent =
     }
   // vnext events
   | WorkflowStreamEvent;
-
-export type WorkflowStreamEvent =
-  | {
-      type: 'workflow-start';
-      payload: {
-        workflowId: string;
-      };
-    }
-  | {
-      type: 'workflow-finish';
-      payload: {
-        workflowStatus: WorkflowRunStatus;
-        output: {
-          usage: {
-            inputTokens: number;
-            outputTokens: number;
-            totalTokens: number;
-          };
-        };
-        metadata: Record<string, any>;
-      };
-    }
-  | {
-      type: 'workflow-canceled';
-      payload: {};
-    }
-  | {
-      type: 'workflow-step-start';
-      id: string;
-      payload: {
-        id: string;
-        stepCallId: string;
-        status: WorkflowStepStatus;
-        output?: Record<string, any>;
-        payload?: Record<string, any>;
-        resumePayload?: Record<string, any>;
-        suspendPayload?: Record<string, any>;
-      };
-    }
-  | {
-      type: 'workflow-step-finish';
-      payload: {
-        id: string;
-        metadata: Record<string, any>;
-      };
-    }
-  | {
-      type: 'workflow-step-suspended';
-      payload: {
-        id: string;
-        status: WorkflowStepStatus;
-        output?: Record<string, any>;
-        payload?: Record<string, any>;
-        resumePayload?: Record<string, any>;
-        suspendPayload?: Record<string, any>;
-      };
-    }
-  | {
-      type: 'workflow-step-waiting';
-      payload: {
-        id: string;
-        payload: Record<string, any>;
-        startedAt: number;
-        status: WorkflowStepStatus;
-      };
-    }
-  | {
-      type: 'workflow-step-result';
-      payload: {
-        id: string;
-        stepCallId: string;
-        status: WorkflowStepStatus;
-        output?: Record<string, any>;
-        payload?: Record<string, any>;
-        resumePayload?: Record<string, any>;
-        suspendPayload?: Record<string, any>;
-      };
-    };
 
 export type WorkflowRunStatus = 'running' | 'success' | 'failed' | 'suspended' | 'waiting' | 'pending' | 'canceled';
